Move the contents of index.tpl to the userFramework specific (meta)template, as standalone has different needs than the others (Responsibility for outputting the <html> and <head>, etc.)
This replaces the metaTpl stuff with going to the userFramework specific template directly (as opposed to indirectly via index.tpl.)

lobo said that the only reason to continue using a lower case template is problems changing the case of filenames on case insensitive systems. If the userFramework template files drupal.tpl/joomla.tpl/standalone.tpl ever move to another folder, it might be a good idea to upper case the first letter to match userFramework.

Fix CRM-2495.

Also, hello everyone. :)